Update crosstool apollo/kronos

+++ make/crosstool.mk
@@ -58,32 +58,20 @@ $(CROSS_DIR)/bin/$(TARGET)-gcc: $(ARCHIVE)/crosstool-ng-$(CROSSTOOL_NG_VER).tar.
 		./configure --enable-local; MAKELEVEL=0 make; chmod 0755 ct-ng; \
 		./ct-ng oldconfig; \
 		./ct-ng build
-	ln -sf sys-root/lib $(CROSS_DIR)/$(TARGET)/
+	test -e $(CROSS_DIR)/$(TARGET)/lib || ln -sf sys-root/lib $(CROSS_DIR)/$(TARGET)/
 
 else ## ifeq ($(PLATFORM), nevis)
 ## apollo / kronos crosstool ###############################################
 
-ifeq ($(UCLIBC_BUILD), 1)
-ifeq ($(USE_UCLIBC_NG), 1)
-## build apollo/kronos uClibc-ng
-CT_NG_CONFIG = $(PATCHES)/ct-ng-1.20/ct-ng-1.20.0-3.config
-else
-## build apollo/kronos uClibc
-CT_NG_CONFIG = $(PATCHES)/ct-ng-1.20/ct-ng-1.21.0-2.config
-endif ## ($(USE_UCLIBC_NG), 1)
-else
-## build apollo/kronos eglibc
-CT_NG_CONFIG = $(PATCHES)/ct-ng-1.20/ct-ng-1.20.0-1-glibc.config
-endif ## ifeq ($(UCLIBC_BUILD), 1)
-
+CT_NG_CONFIG      = $(PATCHES)/ct-ng-1.20/ct-ng-1.22.0-1.config
 CUSTOM_KERNEL_VER = cst-kernel_cst_3.10_2015-10-17_0818_6b9df41
 CUSTOM_KERNEL     = $(ARCHIVE)/linux-$(CUSTOM_KERNEL_VER).tar.xz
-CUSTOM_GCC_VER    = linaro-5.2-2015.11
+CUSTOM_GCC_VER    = linaro-5.2-2015.11-2
 CUSTOM_GCC        = $(ARCHIVE)/gcc-$(CUSTOM_GCC_VER).tar.xz
 
 crosstool: $(CROSS_DIR)/bin/$(TARGET)-gcc
 
-$(CROSS_DIR)/bin/$(TARGET)-gcc: $(ARCHIVE)/crosstool-ng-$(CROSSTOOL_NG_VER).tar.xz $(ARCHIVE)/$(CUSTOM_KERNEL_VER).tar.xz
+$(CROSS_DIR)/bin/$(TARGET)-gcc: $(ARCHIVE)/crosstool-ng-$(CROSSTOOL_NG_VER).tar.xz $(ARCHIVE)/$(CUSTOM_KERNEL_VER).tar.xz $(ARCHIVE)/gcc-$(CUSTOM_GCC_VER).tar.xz
 	make $(BUILD_TMP)
 	if [ ! -e $(CROSS_DIR) ]; then \
 		mkdir -p $(CROSS_DIR); \
@@ -93,15 +81,15 @@ $(CROSS_DIR)/bin/$(TARGET)-gcc: $(ARCHIVE)/crosstool-ng-$(CROSSTOOL_NG_VER).tar.
 	mkdir -p $(BUILD_TMP)/crosstool-ng-$(CROSSTOOL_NG_VER)/targets/src
 	# Don't patch a custom source from ct-ng, it's custom!
 	cd $(BUILD_TMP)/crosstool-ng-$(CROSSTOOL_NG_VER)/targets; \
-		tar -xf $(PATCHES)/ct-ng-1.20/libstdc++_configure_patch.tar.xz; \
-		cd src; \
-		touch .gcc-$(CUSTOM_GCC_VER).extracting; \
-		tar -xf $(CUSTOM_GCC); \
-		rm -f .gcc-$(CUSTOM_GCC_VER).extracting; \
-		touch .gcc-$(CUSTOM_GCC_VER).extracted; \
-		cd gcc-$(CUSTOM_GCC_VER); \
-		patch -p1 < $(BUILD_TMP)/crosstool-ng-$(CROSSTOOL_NG_VER)/targets/patches/gcc/$(CUSTOM_GCC_VER)/900-libstdc++_configure.patch; \
-		cd ..; \
+		tar -xf $(PATCHES)/ct-ng-1.20/libstdc++_configure_patch.tar.xz && \
+		cd src && \
+		touch .gcc-$(CUSTOM_GCC_VER).extracting && \
+		tar -xf $(CUSTOM_GCC) && \
+		rm -f .gcc-$(CUSTOM_GCC_VER).extracting && \
+		touch .gcc-$(CUSTOM_GCC_VER).extracted && \
+		cd gcc-$(CUSTOM_GCC_VER) && \
+		patch -p1 < $(BUILD_TMP)/crosstool-ng-$(CROSSTOOL_NG_VER)/targets/patches/gcc/$(CUSTOM_GCC_VER)/900-libstdc++_configure.patch && \
+		cd .. && \
 		touch .gcc-$(CUSTOM_GCC_VER).patched
 	set -e; unset CONFIG_SITE; cd $(BUILD_TMP)/crosstool-ng-$(CROSSTOOL_NG_VER); \
 		\
@@ -114,28 +102,20 @@ $(CROSS_DIR)/bin/$(TARGET)-gcc: $(ARCHIVE)/crosstool-ng-$(CROSSTOOL_NG_VER).tar.
 		test $$NUM_CPUS -lt 1 && NUM_CPUS=1; \
 		sed -i "s@^CT_PARALLEL_JOBS=.*@CT_PARALLEL_JOBS=$$NUM_CPUS@" .config; \
 		\
-		if [ "$(UCLIBC_BUILD)" = "1" ]; then \
-			cp $(PATCHES)/ct-ng-1.20/900-pull-socket_type-h-from-eglibc.patch patches/uClibc/0.9.33.2; \
-			cp $(PATCHES)/ct-ng-1.20/901-gettimeofday.c-use-the-same-type-as-in-header.patch patches/uClibc/0.9.33.2; \
-			if [ "$(USE_UCLIBC_NG)" = "1" ]; then \
-				export CST_UCLIBC_CONFIG="$(PATCHES)/ct-ng-1.20/ct-ng-uClibc-ng-1.0.1.config"; \
-			else \
-				export CST_UCLIBC_CONFIG="$(PATCHES)/ct-ng-1.20/ct-ng-uClibc-0.9.33.2.config"; \
-			fi;\
-		fi; \
+		cp $(PATCHES)/ct-ng-1.20/900-pull-socket_type-h-from-eglibc.patch patches/uClibc/0.9.33.2; \
+		cp $(PATCHES)/ct-ng-1.20/901-gettimeofday.c-use-the-same-type-as-in-header.patch patches/uClibc/0.9.33.2; \
+		export CST_UCLIBC_CONFIG="$(PATCHES)/ct-ng-1.20/ct-ng-uClibc-0.9.33.2.config"; \
 		\
 		export CST_BASE_DIR=$(BASE_DIR); \
 		export CST_CUSTOM_KERNEL=$(CUSTOM_KERNEL); \
 		export CST_CUSTOM_GCC=$(CUSTOM_GCC); \
+		export CST_CUSTOM_GCC_VER=$(CUSTOM_GCC_VER); \
 		export LD_LIBRARY_PATH=; \
-		if [ "$(USE_UCLIBC_NG)" = "1" ]; then \
-			export CST_CUSTOM_UCLIBC=$(ARCHIVE)/uClibc-ng-1.0.1.tar.xz; \
-		fi; \
 		test -f ./configure || ./bootstrap && \
 		./configure --enable-local; MAKELEVEL=0 make; chmod 0755 ct-ng; \
 		./ct-ng oldconfig; \
 		./ct-ng build
-	ln -sf sys-root/lib $(CROSS_DIR)/$(TARGET)/
+	test -e $(CROSS_DIR)/$(TARGET)/lib || ln -sf sys-root/lib $(CROSS_DIR)/$(TARGET)/
 
 ############################################
